+++ controllers/GENERAL/team.controller.js
@@ -90,13 +90,15 @@ exports.createTeam = catchAsync(async (req, res, next) => {
   if (filteredBody.users && filteredBody.users.length)
     filteredBody.users = filterTeamUsersID(filteredBody.users);
 
-  const foundTeam = await Team.findOne({
-    name: filteredBody.name,
-    status: 'Deleted',
-    _tenantId: req.user._tenantId,
-  });
+  if (filteredBody.name) {
+    const foundTeam = await Team.findOne({
+      name: filteredBody.name,
+      status: 'Deleted',
+      _tenantId: req.user._tenantId,
+    });
 
-  if (foundTeam) await Team.findByIdAndDelete(foundTeam._id);
+    if (foundTeam) await Team.findByIdAndDelete(foundTeam._id);
+  }
 
   const team = await Team.create(filteredBody);
 
@@ -175,13 +177,15 @@ exports.updateTeam = catchAsync(async (req, res, next) => {
   if (filteredBody.users && filteredBody.users.length)
     filteredBody.users = filterTeamUsersID(filteredBody.users);
 
-  const foundTeam = await Team.findOne({
-    name: filteredBody.name || '',
-    status: 'Deleted',
-    _tenantId: req.user._tenantId,
-  });
+  if (filteredBody.name) {
+    const foundTeam = await Team.findOne({
+      name: filteredBody.name,
+      status: 'Deleted',
+      _tenantId: req.user._tenantId,
+    });
 
-  if (foundTeam) await Team.findByIdAndDelete(foundTeam._id);
+    if (foundTeam) await Team.findByIdAndDelete(foundTeam._id);
+  }
 
   const updatedTeam = await Team.findByIdAndUpdate(team._id, filteredBody, {
     new: true,
